large model and blinking led with timer working

+++ payload/drivers/msp430fr5xx_6xx/inc/hw_memmap.h
@@ -52,8 +52,8 @@
 // SUCCESS and FAILURE for API return value
 //
 //*****************************************************************************
-#define STATUS_SUCCESS  0x01
-#define STATUS_FAIL     0x00
+#define STATUS_SUCCESS 0x01
+#define STATUS_FAIL 0x00
 
 //*****************************************************************************
 //
@@ -67,11 +67,11 @@
 // Macros for hardware access
 //
 //*****************************************************************************
-#define HWREG32(x)                                                              \
-    (*((volatile uint32_t *)((uint16_t)x)))
-#define HWREG16(x)                                                             \
-    (*((volatile uint16_t *)((uint16_t)x)))
-#define HWREG8(x)                                                             \
-    (*((volatile uint8_t *)((uint16_t)x)))
+#define HWREG32(x) \
+    (*((volatile uint32_t *)((uintptr_t)x)))
+#define HWREG16(x) \
+    (*((volatile uint16_t *)((uintptr_t)x)))
+#define HWREG8(x) \
+    (*((volatile uint8_t *)((uintptr_t)x)))
 
 #endif // #ifndef __HW_MEMMAP__
